+++ hpipy/datasets/__init__.py
@@ -1,0 +1,5 @@
+"""Datasets module."""
+
+from ._base import load_ex_sales, load_seattle_sales
+
+__all__ = ["load_ex_sales", "load_seattle_sales"]