+++ hpipy/datasets/_base.py
@@ -1,0 +1,23 @@
+import importlib.resources as pkg_resources
+
+import pandas as pd
+
+
+def load_ex_sales():
+    """Load example home sales dataset.
+
+    Returns:
+        pandas.DataFrame: A DataFrame with house saless.
+    """
+    with pkg_resources.files("hpipy.datasets.data").joinpath("ex_sales.csv").open("r") as f:
+        return pd.read_csv(f, parse_dates=["sale_date"])
+
+
+def load_seattle_sales():
+    """Load Seattle home sales dataset.
+
+    Returns:
+        pandas.DataFrame: A DataFrame with house sales.
+    """
+    with pkg_resources.files("hpipy.datasets.data").joinpath("seattle_sales.csv").open("r") as f:
+        return pd.read_csv(f, parse_dates=["sale_date"])